Integrated PlaceholderValidator to MailForm and MailSettingsForm
