+++ src/pretix/plugins/sendmail/forms.py
@@ -1,40 +1,9 @@
-import re
 from django import forms
-from django.core.exceptions import ValidationError
-from django.core.validators import BaseValidator
 from django.utils.translation import ugettext_lazy as _
 from i18nfield.forms import I18nFormField, I18nTextarea, I18nTextInput
 
+from pretix.base.forms import PlaceholderValidator
 from pretix.base.models import Order
-
-
-class PlaceholderValidator(BaseValidator):
-    """
-    Takes list of allowed placeholders,
-    validates form field by checking for placeholders,
-    which are not presented in taken list.
-    """
-
-    def __init__(self, limit_value):
-        super().__init__(limit_value)
-        self.limit_value = limit_value
-
-    def __call__(self, value):
-        data_placeholders = list(re.findall(r'({[\w\s]*})', str(value), re.X))
-
-        invalid_placeholders = []
-        for placeholder in data_placeholders:
-            if placeholder not in self.limit_value:
-                invalid_placeholders.append(placeholder)
-        if invalid_placeholders:
-            raise ValidationError(
-                _('Invalid placeholder(s): %(value)s'),
-                code='invalid',
-                params={'value': ", ".join(invalid_placeholders)},
-            )
-
-    def clean(self, x):
-        return x
 
 
 class MailForm(forms.Form):
